refactor: 탭 관련 서비스 패키지 변경
